refactor: creating a separate file for processing_and_reuse_operations
