+++ back/src/common/constants/index.ts
@@ -1,4 +1,5 @@
 export * from "./PROCESSING_OPERATIONS";
+export * from "./PROCESSING_AND_REUSE_OPERATIONS";
 export * from "./WASTES";
 export * from "./DASRI_CONSTANTS";
 export * from "./ROAD_CONTROL";
